requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
jinja2==3.1.3
python-dotenv==1.0.1
aiofiles==23.2.1
